+++ src/packages/control/Geolocation/index.ts
@@ -1,0 +1,5 @@
+import { withInstall } from '../../../utils/with-install'
+import Geolocation from './Geolocation.vue'
+
+export const ElAmapControlGeolocation = withInstall(Geolocation)
+export default ElAmapControlGeolocation